domain/src/test/kotlin/mega/privacy/android/domain/usecase/camerauploads/SetUploadFileNamesKeptUseCaseTest.kt
package mega.privacy.android.domain.usecase.camerauploads

import kotlinx.coroutines.ExperimentalCoroutinesApi
import kotlinx.coroutines.test.runTest
import mega.privacy.android.domain.repository.CameraUploadsRepository
import org.junit.jupiter.api.BeforeAll
import org.junit.jupiter.api.BeforeEach
import org.junit.jupiter.api.TestInstance
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.ValueSource
import org.mockito.kotlin.mock
import org.mockito.kotlin.reset
import org.mockito.kotlin.times
import org.mockito.kotlin.verify

/**
 * Test class for [SetUploadFileNamesKeptUseCase]
 */
@ExperimentalCoroutinesApi
@TestInstance(TestInstance.Lifecycle.PER_CLASS)
class SetUploadFileNamesKeptUseCaseTest {

    private lateinit var underTest: SetUploadFileNamesKeptUseCase

    private val cameraUploadsRepository = mock<CameraUploadsRepository>()

    @BeforeAll
    fun setUp() {
        underTest = SetUploadFileNamesKeptUseCase(
            cameraUploadsRepository = cameraUploadsRepository,
        )
    }

    @BeforeEach
    fun resetMocks() {
        reset(cameraUploadsRepository)
    }

    @ParameterizedTest(name = "keep file names: {0}")
    @ValueSource(booleans = [true, false])
    fun `test that the file names for uploads are handled`(keepFileNames: Boolean) = runTest {
        underTest(keepFileNames)

        verify(cameraUploadsRepository, times(1)).setUploadFileNamesKept(keepFileNames)
    }
}